requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
ta==0.10.2
matplotlib==3.7.2
plotly==5.17.0
python-telegram-bot==20.7
APScheduler==3.10.4
python-dotenv==1.0.0
ccxt==4.1.77
websocket-client==1.6.4
aiohttp==3.9.1
asyncio-mqtt==0.13.0
flask==3.0.0
flask-cors==4.0.0

